added basic logic to script, now checking for variables and creating env file when run for the first time

+++ shellscripts/download_gitlab_data.sh
@@ -1,9 +1,15 @@
-# paste your token here
-gitlab_token=''
+# get gitlab token
+export_script="export_env_variables.sh"
+source $export_script
+if [ -z $GITLAB_SOURCE_TOKEN ]
+then
+    echo "gitlab token not provided"
+    exit 1
+fi
 gitlab_project_id='262'
 gitlab_api_base_url='https://gitlab.oeaw.ac.at/api/v4/projects/'
 gitlab_api_request='/repository/archive?'
-output_archive="./flugblatter.tar.gz"
+output_archive="./flugblätter.tar.gz"
 custom_output_dir="./todesurteile_master/"
 # rm old data / create target dir
 if [ -d $custom_output_dir ]; then rm -r $custom_output_dir; fi
@@ -11,7 +17,7 @@ mkdir $custom_output_dir
 # to get an overview of what you could do with API/Token, use:
 # curl "${gitlab_api_base_url}?private_token=${gitlab_token}"
 # get the data via api
-curl "${gitlab_api_base_url}${gitlab_project_id}${gitlab_api_request}private_token=${gitlab_token}" --output $output_archive
+curl "${gitlab_api_base_url}${gitlab_project_id}${gitlab_api_request}private_token=${GITLAB_SOURCE_TOKEN}" --output $output_archive
 # find out, how the api named the parent folder in the archive
 top_level_folder_name="./"`tar -tzf $output_archive | head -1 | cut -f1 -d"/"`
 # unzip the archive
